chore: move contract interfaces to dedicated folder
