+++ packages/mock-utils/src/fhevm/contracts/interfaces/ZamaFheDecryptionOracle.itf.ts
