+++ packages/mock-utils/src/fhevm/coprocessor/MockCoprocessor.ts
@@ -4,8 +4,8 @@ import { multiSignEIP712 } from "../../ethers/eip712.js";
 import { assertIsBytes32String } from "../../utils/bytes.js";
 import { assertFhevm } from "../../utils/error.js";
 import { removePrefix } from "../../utils/string.js";
-import { FHEVMExecutorPartialInterface } from "../contracts/FHEVMExecutor.itf.js";
 import { InputVerifier } from "../contracts/InputVerifier.js";
+import { FHEVMExecutorPartialInterface } from "../contracts/interfaces/FHEVMExecutor.itf.js";
 import type { FhevmDB, FhevmDBHandleMetadata } from "../db/FhevmDB.js";
 import type { RelayerV1InputProofResponse } from "../relayer/payloads.js";
 import type { Coprocessor } from "./Coprocessor.js";
